changed the data type of the setter from int to long

+++ backend/src/main/java/ca/mcgill/ecse321/Mar1HotelSystem/model/GeneralUser.java
@@ -73,7 +73,7 @@ public class GeneralUser {
     }
 
     // Method to set phoneNumber, returns true if phoneNumber set
-    public boolean setPhoneNumber(int phoneNumber) {
+    public boolean setPhoneNumber(long phoneNumber) {
         this.phoneNumber = phoneNumber;
         return true;
     }
